8290 Use one ReaderConfigProvider for the app’s lifetime
Once the CardReaderConfigProvider is set on the Terminal, it can’t be changed. This means we need to keep the same one around for the lifetime of the app in order to change it when the user logs out and back in again.

Failing to do this results in serious errors when using a second account to take IPP transactions.
