+++ WooCommerce/Classes/Yosemite/AuthenticatedState.swift
@@ -75,7 +75,8 @@ class AuthenticatedState: StoresManagerState {
             CardPresentPaymentStore(dispatcher: dispatcher,
                                     storageManager: storageManager,
                                     network: network,
-                                    cardReaderService: ServiceLocator.cardReaderService),
+                                    cardReaderService: ServiceLocator.cardReaderService,
+                                    cardReaderConfigProvider: ServiceLocator.cardReaderConfigProvider),
             ReceiptStore(dispatcher: dispatcher,
                          storageManager: storageManager,
                          network: network,
